--- instrumentation/external-annotations/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/extannotations/TraceAdvice.java
@@ -1,38 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.extannotations;
-
-import static io.opentelemetry.javaagent.instrumentation.extannotations.TraceAnnotationTracer.tracer;
-
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
-import java.lang.reflect.Method;
-import net.bytebuddy.asm.Advice;
-
-public class TraceAdvice {
-
-  @Advice.OnMethodEnter(suppress = Throwable.class)
-  public static void onEnter(
-      @Advice.Origin Method method,
-      @Advice.Local("otelContext") Context context,
-      @Advice.Local("otelScope") Scope scope) {
-    context = tracer().startSpan(method);
-    scope = context.makeCurrent();
-  }
-
-  @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-  public static void stopSpan(
-      @Advice.Local("otelContext") Context context,
-      @Advice.Local("otelScope") Scope scope,
-      @Advice.Thrown Throwable throwable) {
-    scope.close();
-    if (throwable != null) {
-      tracer().endExceptionally(context, throwable);
-    } else {
-      tracer().end(context);
-    }
-  }
-}
