+++ instrumentation/play-ws/play-ws-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/playws/v1_0/PlayWsInstrumentationModule.java
@@ -30,7 +30,8 @@ public class PlayWsInstrumentationModule extends InstrumentationModule {
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new AsyncHttpClientInstrumentation(ClientAdvice.class.getName()));
+    return singletonList(new AsyncHttpClientInstrumentation(
+        PlayWsInstrumentationModule.class.getName() + "$ClientAdvice"));
   }
 
   public static class ClientAdvice {
